Corretti errori trovati nella revisione
