+++ frontend/lib/domain/models/material/resource.dart
