+++ frontend/lib/domain/models/material/resource_type.dart
