+++ frontend/test/data/dtos/resource_dto_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mvp_app_protegge_e_trasforma/data/dtos/resource_dto.dart';
-import 'package:mvp_app_protegge_e_trasforma/domain/resource_type.dart';
+import 'package:mvp_app_protegge_e_trasforma/domain/models/material/resource_type.dart';
 
 void main() {
   group('ResourceDTO Unit Test', () {
